[ADD] Added eventDetails component and modified userprofiles
