+++ package.json
@@ -9,11 +9,11 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "core-js": "^3.3.2",
+    "core-js": "^3.4.0",
     "node-sass": "^4.13.0",
     "sass-loader": "^8.0.0",
     "ts-loader": "^6.2.1",
-    "typescript": "^3.6.4",
+    "typescript": "^3.7.2",
     "vue": "^2.6.10",
     "vue-router": "^3.1.3"
   },
