+++ src/App.vue
@@ -1,8 +1,5 @@
 <template>
   <div>
-    <div>
-      <router-link to="/events">Home</router-link>
-    </div>
     <router-view/>
   </div>
 </template>
