+++ src/pages/Events/EventDetails.vue
@@ -1,0 +1,13 @@
+<template>
+    <div></div>
+</template>
+
+<script>
+export default {
+
+};
+</script>
+
+<style lang="scss" scoped>
+
+</style>
